+++ src/cryptography/utils.py
@@ -18,6 +18,11 @@ PersistentlyDeprecated = DeprecationWarning
 DeprecatedIn19 = DeprecationWarning
 
 
+def _check_bytes(name, value):
+    if not isinstance(value, bytes):
+        raise TypeError("{0} must be bytes".format(name))
+
+
 def read_only_property(name):
     return property(lambda self: getattr(self, name))